Isolate method functions into its own runtime module
