+++ opal/runtime/boot.js
@@ -760,85 +760,6 @@
     return klass;
   };
 
-  Opal.is_method = function(prop) {
-    return (prop[0] === '$' && prop[1] !== '$');
-  };
-
-  Opal.instance_methods = function(mod) {
-    var processed = Object.create(null), results = [], ancestors = $ancestors(mod);
-
-    for (var i = 0, l = ancestors.length; i < l; i++) {
-      var ancestor = ancestors[i],
-          proto = ancestor.$$prototype;
-
-      if (proto.hasOwnProperty('$$dummy')) {
-        proto = proto.$$define_methods_on;
-      }
-
-      var props = Object.getOwnPropertyNames(proto);
-
-      for (var j = 0, ll = props.length; j < ll; j++) {
-        var prop = props[j];
-
-        if (processed[prop]) {
-          continue;
-        }
-        if (Opal.is_method(prop)) {
-          var method = proto[prop];
-
-          if (!method.$$stub) {
-            var method_name = prop.slice(1);
-            results.push(method_name);
-          }
-        }
-        processed[prop] = true;
-      }
-    }
-
-    return results;
-  };
-
-  Opal.own_instance_methods = function(mod) {
-    var results = [],
-        proto = mod.$$prototype;
-
-    if (proto.hasOwnProperty('$$dummy')) {
-      proto = proto.$$define_methods_on;
-    }
-
-    var props = Object.getOwnPropertyNames(proto);
-
-    for (var i = 0, length = props.length; i < length; i++) {
-      var prop = props[i];
-
-      if (Opal.is_method(prop)) {
-        var method = proto[prop];
-
-        if (!method.$$stub) {
-          var method_name = prop.slice(1);
-          results.push(method_name);
-        }
-      }
-    }
-
-    return results;
-  };
-
-  Opal.methods = function(obj) {
-    return Opal.instance_methods(obj.$$meta || obj.$$class);
-  };
-
-  Opal.own_methods = function(obj) {
-    return obj.$$meta ? Opal.own_instance_methods(obj.$$meta) : [];
-  };
-
-  Opal.receiver_methods = function(obj) {
-    var mod = Opal.get_singleton_class(obj);
-    var singleton_methods = Opal.own_instance_methods(mod);
-    var instance_methods = Opal.own_instance_methods(mod.$$super);
-    return singleton_methods.concat(instance_methods);
-  };
-
   // Returns an object containing all pairs of names/values
   // for all class variables defined in provided +module+
   // and its ancestors.
@@ -1429,26 +1350,6 @@
   // Methods
   // -------
 
-  // Arity count error dispatcher for methods
-  //
-  // @param actual [Fixnum] number of arguments given to method
-  // @param expected [Fixnum] expected number of arguments
-  // @param object [Object] owner of the method +meth+
-  // @param meth [String] method name that got wrong number of arguments
-  // @raise [ArgumentError]
-  Opal.ac = function(actual, expected, object, meth) {
-    var inspect = '';
-    if (object.$$is_a_module) {
-      inspect += object.$$name + '.';
-    }
-    else {
-      inspect += object.$$class.$$name + '#';
-    }
-    inspect += meth;
-
-    $raise(Opal.ArgumentError, '[' + inspect + '] wrong number of arguments (given ' + actual + ', expected ' + expected + ')');
-  };
-
   // Arity count error dispatcher for blocks
   //
   // @param actual [Fixnum] number of arguments given to block
@@ -1715,245 +1616,6 @@
     $apply_blockopts(block, blockopts);
 
     return block;
-  };
-
-  // Used to define methods on an object. This is a helper method, used by the
-  // compiled source to define methods on special case objects when the compiler
-  // can not determine the destination object, or the object is a Module
-  // instance. This can get called by `Module#define_method` as well.
-  //
-  // ## Modules
-  //
-  // Any method defined on a module will come through this runtime helper.
-  // The method is added to the module body, and the owner of the method is
-  // set to be the module itself. This is used later when choosing which
-  // method should show on a class if more than 1 included modules define
-  // the same method. Finally, if the module is in `module_function` mode,
-  // then the method is also defined onto the module itself.
-  //
-  // ## Classes
-  //
-  // This helper will only be called for classes when a method is being
-  // defined indirectly; either through `Module#define_method`, or by a
-  // literal `def` method inside an `instance_eval` or `class_eval` body. In
-  // either case, the method is simply added to the class' prototype. A special
-  // exception exists for `BasicObject` and `Object`. These two classes are
-  // special because they are used in toll-free bridged classes. In each of
-  // these two cases, extra work is required to define the methods on toll-free
-  // bridged class' prototypes as well.
-  //
-  // ## Objects
-  //
-  // If a simple ruby object is the object, then the method is simply just
-  // defined on the object as a singleton method. This would be the case when
-  // a method is defined inside an `instance_eval` block.
-  //
-  // @param obj  [Object, Class] the actual obj to define method for
-  // @param jsid [String] the JavaScript friendly method name (e.g. '$foo')
-  // @param body [JS.Function] the literal JavaScript function used as method
-  // @param blockopts [Object, Number] optional properties to set on the body
-  // @return [null]
-  //
-  Opal.def = function(obj, jsid, body, blockopts) {
-    $apply_blockopts(body, blockopts);
-
-    // Special case for a method definition in the
-    // top-level namespace
-    if (obj === Opal.top) {
-      return Opal.defn(Opal.Object, jsid, body);
-    }
-    // if instance_eval is invoked on a module/class, it sets inst_eval_mod
-    else if (!obj.$$eval && obj.$$is_a_module) {
-      return Opal.defn(obj, jsid, body);
-    }
-    else {
-      return Opal.defs(obj, jsid, body);
-    }
-  };
-
-  // Define method on a module or class (see Opal.def).
-  Opal.defn = function(module, jsid, body) {
-    $deny_frozen_access(module);
-
-    body.displayName = jsid;
-    body.$$owner = module;
-
-    var name = jsid.substr(1);
-
-    var proto = module.$$prototype;
-    if (proto.hasOwnProperty('$$dummy')) {
-      proto = proto.$$define_methods_on;
-    }
-    $prop(proto, jsid, body);
-
-    if (module.$$is_module) {
-      if (module.$$module_function) {
-        Opal.defs(module, jsid, body)
-      }
-
-      for (var i = 0, iclasses = module.$$iclasses, length = iclasses.length; i < length; i++) {
-        var iclass = iclasses[i];
-        $prop(iclass, jsid, body);
-      }
-    }
-
-    var singleton_of = module.$$singleton_of;
-    if (module.$method_added && !module.$method_added.$$stub && !singleton_of) {
-      module.$method_added(name);
-    }
-    else if (singleton_of && singleton_of.$singleton_method_added && !singleton_of.$singleton_method_added.$$stub) {
-      singleton_of.$singleton_method_added(name);
-    }
-
-    return name;
-  };
-
-  // Define a singleton method on the given object (see Opal.def).
-  Opal.defs = function(obj, jsid, body, blockopts) {
-    $apply_blockopts(body, blockopts);
-
-    if (obj.$$is_string || obj.$$is_number) {
-      $raise(Opal.TypeError, "can't define singleton");
-    }
-    return Opal.defn(Opal.get_singleton_class(obj), jsid, body);
-  };
-
-  // Since JavaScript has no concept of modules, we create proxy classes
-  // called `iclasses` that store copies of methods loaded. We need to
-  // update them if we remove a method.
-  function remove_method_from_iclasses(obj, jsid) {
-    if (obj.$$is_module) {
-      for (var i = 0, iclasses = obj.$$iclasses, length = iclasses.length; i < length; i++) {
-        var iclass = iclasses[i];
-        delete iclass[jsid];
-      }
-    }
-  }
-
-  // Called from #remove_method.
-  Opal.rdef = function(obj, jsid) {
-    if (!$has_own(obj.$$prototype, jsid)) {
-      $raise(Opal.NameError, "method '" + jsid.substr(1) + "' not defined in " + obj.$name());
-    }
-
-    delete obj.$$prototype[jsid];
-
-    remove_method_from_iclasses(obj, jsid);
-
-    if (obj.$$is_singleton) {
-      if (obj.$$prototype.$singleton_method_removed && !obj.$$prototype.$singleton_method_removed.$$stub) {
-        obj.$$prototype.$singleton_method_removed(jsid.substr(1));
-      }
-    }
-    else {
-      if (obj.$method_removed && !obj.$method_removed.$$stub) {
-        obj.$method_removed(jsid.substr(1));
-      }
-    }
-  };
-
-  // Called from #undef_method.
-  Opal.udef = function(obj, jsid) {
-    if (!obj.$$prototype[jsid] || obj.$$prototype[jsid].$$stub) {
-      $raise(Opal.NameError, "method '" + jsid.substr(1) + "' not defined in " + obj.$name());
-    }
-
-    Opal.add_stub_for(obj.$$prototype, jsid);
-
-    remove_method_from_iclasses(obj, jsid);
-
-    if (obj.$$is_singleton) {
-      if (obj.$$prototype.$singleton_method_undefined && !obj.$$prototype.$singleton_method_undefined.$$stub) {
-        obj.$$prototype.$singleton_method_undefined(jsid.substr(1));
-      }
-    }
-    else {
-      if (obj.$method_undefined && !obj.$method_undefined.$$stub) {
-        obj.$method_undefined(jsid.substr(1));
-      }
-    }
-  };
-
-  function is_method_body(body) {
-    return (typeof(body) === "function" && !body.$$stub);
-  }
-
-  Opal.alias = function(obj, name, old) {
-    var id     = $jsid(name),
-        old_id = $jsid(old),
-        body,
-        alias;
-
-    // Aliasing on main means aliasing on Object...
-    if (typeof obj.$$prototype === 'undefined') {
-      obj = Opal.Object;
-    }
-
-    body = obj.$$prototype[old_id];
-
-    // When running inside #instance_eval the alias refers to class methods.
-    if (obj.$$eval) {
-      return Opal.alias(Opal.get_singleton_class(obj), name, old);
-    }
-
-    if (!is_method_body(body)) {
-      var ancestor = obj.$$super;
-
-      while (typeof(body) !== "function" && ancestor) {
-        body     = ancestor[old_id];
-        ancestor = ancestor.$$super;
-      }
-
-      if (!is_method_body(body) && obj.$$is_module) {
-        // try to look into Object
-        body = Opal.Object.$$prototype[old_id]
-      }
-
-      if (!is_method_body(body)) {
-        $raise(Opal.NameError, "undefined method `" + old + "' for class `" + obj.$name() + "'")
-      }
-    }
-
-    // If the body is itself an alias use the original body
-    // to keep the max depth at 1.
-    if (body.$$alias_of) body = body.$$alias_of;
-
-    // We need a wrapper because otherwise properties
-    // would be overwritten on the original body.
-    alias = Opal.wrap_method_body(body);
-
-    // Try to make the browser pick the right name
-    alias.displayName  = name;
-    alias.$$alias_of   = body;
-    alias.$$alias_name = name;
-
-    Opal.defn(obj, id, alias);
-
-    return obj;
-  };
-
-  Opal.wrap_method_body = function(body) {
-    var wrapped = function() {
-      var block = wrapped.$$p;
-
-      wrapped.$$p = null;
-
-      return Opal.send(this, body, arguments, block);
-    };
-
-    // Assign the 'length' value with defineProperty because
-    // in strict mode the property is not writable.
-    // It doesn't work in older browsers (like Chrome 38), where
-    // an exception is thrown breaking Opal altogether.
-    try {
-      Object.defineProperty(wrapped, 'length', { value: body.length });
-    } catch {}
-
-    wrapped.$$arity           = body.$$arity == null ? body.length : body.$$arity;
-    wrapped.$$parameters      = body.$$parameters;
-    wrapped.$$source_location = body.$$source_location;
-
-    return wrapped;
   };
 
   Opal.alias_gvar = function(new_name, old_name) {
